When there are changes on the application page, the number and status of the like also change in the reviews tab.
